break down npcGraphic types
